feat: add safety by activating strictNullChecks (#169)
* add safety by activating strictNullChecks

* modification retries tests

+++ src/test/hooks.ts
@@ -3,7 +3,7 @@ import { Database } from '../db/conn';
 export const mochaHooks = {
   afterEach: async function () {
     const db = await Database.getInstance();
-    if (db.namespace === 'grindery-test-server') {
+    if (db && db.namespace === 'grindery-test-server') {
       await db.dropDatabase();
     }
   },
